requirements.txt
python-dotenv==1.0.0
openai==1.12.0
langchain==0.1.12
langchain-community==0.0.27
chromadb==0.4.24
tiktoken==0.6.0
sentence-transformers==2.5.1
streamlit==1.31.1
pillow==10.2.0  # 이미지 처리 
langchain-openai==0.0.2  # 새로 추가 